requirements.txt
cryptography==37.0.4
PyJWT==2.4.0
requests==2.27.1